+++ Flight.py
@@ -10,8 +10,8 @@ class Flight(Base):
 
     id = Column(BigInteger(), primary_key=True, autoincrement=True)
     airline_Company_Id= Column(BigInteger, ForeignKey('airline_companies.id', ondelete='CASCADE'))
-    origin_Country_id= Column(Integer, ForeignKey ('countries.id', ondelete='CASCADE'), nullable=False)
-    destination_Country_id=Column(Integer, ForeignKey ('countries.id', ondelete='CASCADE'), nullable=False)
+    origin_Country_id= Column(BigInteger, ForeignKey ('countries.id', ondelete='CASCADE'), nullable=False)
+    destination_Country_id=Column(BigInteger, ForeignKey ('countries.id', ondelete='CASCADE'), nullable=False)
     departure_Time=Column(DateTime()) #add nullable=False
     landing_Time=Column(DateTime()) #nullable=False
     remaining_Tickets=Column(Integer(), nullable=False)
